cleanup cont
